feat(console): add FrameworkIntegration, ObservabilityFeatures,
